# docs/智能内容运营平台/设计/设计细节/openapi.yaml
openapi: 3.0.3
info:
  title: Agentic Content Studio API
  version: 1.0.0
  description: 智能内容运营平台 API，覆盖账号、选题、内容、Agent、审核、发布和数据复盘主流程。
servers:
  - url: http://localhost:3002/api
    description: 本地 API 服务
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        default: 20
...
  schemas:
...
    Platform:
      type: string
      enum: [WECHAT, XIAOHONGSHU, DOUYIN, VIDEO_CHANNEL, BILIBILI, ZHIHU, OTHER]
    ContentStatus:
      type: string
      enum: [DRAFT, PENDING_GENERATE, GENERATING, PENDING_REVIEW, REJECTED, APPROVED, PENDING_PUBLISH, PUBLISHING, PUBLISHED, FAILED, REVIEWED, ARCHIVED]
...
    MaterialType:
      type: string
      enum: [IMAGE, VIDEO, AUDIO, FILE]
    MaterialRole:
      type: string
      enum: [COVER, BODY, ATTACHMENT]
...
    TopicCreate:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
        targetPlatforms:
          type: array
          items:
            $ref: '#/components/schemas/Platform'
        source:
          type: string
          default: manual
...
    MaterialInput:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/MaterialType'
        role:
          $ref: '#/components/schemas/MaterialRole'
        name:
          type: string
        url:
          type: string
        localPath:
          type: string
        source:
          type: string
        meta:
          type: object
          additionalProperties: true
...
paths:
...
  /topics:
    get:
      tags: [Topics]
      summary: 选题列表
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ContentStatus'
        - name: ownerId
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: OK
    post:
      tags: [Topics]
      summary: 创建选题
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TopicCreate'
      responses:
        '200':
          description: OK
  /topics/{id}:
    get:
      tags: [Topics]
      summary: 选题详情
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: OK
    patch:
      tags: [Topics]
      summary: 更新选题
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: OK
    delete:
      tags: [Topics]
      summary: 删除选题
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: OK
...
  /contents/{id}/materials:
    get:
      tags: [Materials]
      summary: 内容绑定素材列表
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: OK
    post:
      tags: [Materials]
      summary: 给内容新增素材
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaterialInput'
      responses:
        '200':
          description: OK
...
  /materials:
    get:
      tags: [Materials]
      summary: 素材列表
      parameters:
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/MaterialType'
        - name: role
          in: query
          schema:
            $ref: '#/components/schemas/MaterialRole'
      responses:
        '200':
          description: OK
  /materials/{id}:
    get:
      tags: [Materials]
      summary: 素材详情
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: OK
    patch:
      tags: [Materials]
      summary: 更新素材
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaterialInput'
      responses:
        '200':
          description: OK
    delete:
      tags: [Materials]
      summary: 删除素材
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: OK
...